Changes to tooling with disturbing effects on python execution.
At present running `tox -q` with a poetry environment selected
but not activated causes tests to be run with that environment's
Python, but coverage is correctly recorded.

After `poetry env use system` running the same command runs the
tests using the correct (tox environment) Python, but coverage
is not recorded at all.

The command `poetry run tox -q` also runs the test with the
correct interpreter with no coverage.